+++ archivebox/plugins/chrome/on_Crawl__01_chrome_install.py
