+++ archivebox/plugins/search_backend_ripgrep/on_Crawl__00_ripgrep_install.py
