+++ archivebox/plugins/singlefile/on_Crawl__04_singlefile_install.js
